Revert "Update univalue_write.cpp"
This reverts commit e1eb876dfe090a5d3c59a643aa14b8c283457efa.

+++ src/univalue/lib/univalue_write.cpp
@@ -64,7 +64,7 @@ std::string UniValue::write(unsigned int prettyIndent,
 
 static void indentStr(unsigned int prettyIndent, unsigned int indentLevel, std::string& s)
 {
-    s.append(static_cast<size_t>(prettyIndent) * indentLevel, ' ');
+    s.append(prettyIndent * indentLevel, ' ');
 }
 
 // NOLINTNEXTLINE(misc-no-recursion)
